8_ai-optimized-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
transformers==4.44.2
torch==2.7.1
sqlalchemy==2.0.35
fastapi-cache2
redis==5.0.8
pytest==8.3.2
httpx==0.27.2
